+++ senatus/src/extensions/documentation/documentation/1.0.0/full_documentation.json
@@ -14,7 +14,7 @@
       "name": "Apache 2.0",
       "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
     },
-    "x-generation-date": "2026-03-01T17:24:56.432Z"
+    "x-generation-date": "2026-04-14T20:32:10.720Z"
   },
   "x-strapi-config": {
     "plugins": [
