+++ web/src/pages/login/components/LoginFormFooter.tsx
@@ -1,0 +1,19 @@
+import { Link } from 'react-router-dom';
+import { Button } from '@/components/ui/button';
+import { CardFooter } from '@/components/ui/card';
+
+export default function LoginFormFooter({isLoading}: {isLoading: boolean}) {
+  return (
+    <CardFooter className="flex flex-col gap-4">
+      <Button type="submit" className="w-full" disabled={isLoading}>
+        {isLoading ? 'Signing in...' : 'Sign in'}
+      </Button>
+      <p className="text-sm text-center text-muted-foreground">
+        Don't have an account?{' '}
+        <Link to="/register" className="text-primary hover:underline font-medium">
+          Sign up
+        </Link>
+      </p>
+    </CardFooter>
+  );
+}
